fixing warnings refs #12
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@10076 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/netimport/NIImporter_SUMO.cpp
@@ -474,10 +474,10 @@ NIImporter_SUMO::approximateEdgeShape(const EdgeAttrs* edge) {
         Position2D me = firstLane[i];
         Position2D to = firstLane[i+1];
         std::pair<SUMOReal, SUMOReal> offsets = NBEdge::laneOffset(
-                from, me, SUMO_const_laneWidthAndOffset, noLanes-1, 
+                from, me, SUMO_const_laneWidthAndOffset, (unsigned int)noLanes-1, 
                 noLanes, NBEdge::LANESPREAD_RIGHT, false);
         std::pair<SUMOReal, SUMOReal> offsets2 = NBEdge::laneOffset(
-                me, to, SUMO_const_laneWidthAndOffset, noLanes-1, 
+                me, to, SUMO_const_laneWidthAndOffset, (unsigned int)noLanes-1, 
                 noLanes, NBEdge::LANESPREAD_RIGHT, false);
 
         Line2D l1(
